requirements.txt
pyserial>=3.5
numpy>=1.24.0
pandas>=1.5.2
matplotlib>=3.6.2
ipykernel>=6.23.1